requirements.txt
ipywidgets

python-dotenv

google-auth
google-api-python-client

langchain_community
langchainhub
langchain-huggingface
langchain_experimental

langchain_openai
langchain_google_genai
langchain_mistralai
llama-cpp-python

gigachat
yandexcloud

telebot

pdfminer.six
unstructured
python-docx

